requirements.txt
requests==2.31.0
python-dotenv==1.0.0
discord-webhook==1.3.0
APScheduler==3.10.4
pytz==2024.1
